Refactor polls extension implementation and qualified names
