+++ src/main/kotlin/ar/pelotude/ohhsugoi/bot/UtilsExtension.kt
@@ -3,10 +3,6 @@ package ar.pelotude.ohhsugoi.bot
 import ar.pelotude.ohhsugoi.bot.converters.date
 import ar.pelotude.ohhsugoi.bot.converters.postId
 import ar.pelotude.ohhsugoi.db.MangaDatabase
-import ar.pelotude.ohhsugoi.db.Poll
-import ar.pelotude.ohhsugoi.db.PollException
-import ar.pelotude.ohhsugoi.db.PollOption
-import ar.pelotude.ohhsugoi.db.PollUnsuccessfulOperationException
 import ar.pelotude.ohhsugoi.db.PollsDatabase
 import ar.pelotude.ohhsugoi.db.UserData
 import ar.pelotude.ohhsugoi.db.UsersDatabase
@@ -20,78 +16,42 @@ import ar.pelotude.ohhsugoi.db.scheduler.platforms.DiscordWebhookMessage
 import ar.pelotude.ohhsugoi.db.scheduler.platforms.XPost
 import ar.pelotude.ohhsugoi.util.image.asJpgByteArray
 import ar.pelotude.ohhsugoi.util.image.stitch
+import ar.pelotude.ohhsugoi.util.toMentionOn
 import com.kotlindiscord.kord.extensions.checks.hasRole
 import com.kotlindiscord.kord.extensions.commands.Arguments
 import com.kotlindiscord.kord.extensions.commands.application.slash.converters.impl.stringChoice
 import com.kotlindiscord.kord.extensions.commands.application.slash.ephemeralSubCommand
 import com.kotlindiscord.kord.extensions.commands.application.slash.publicSubCommand
 import com.kotlindiscord.kord.extensions.commands.converters.impl.long
-import com.kotlindiscord.kord.extensions.commands.converters.impl.optionalBoolean
 import com.kotlindiscord.kord.extensions.commands.converters.impl.optionalLong
 import com.kotlindiscord.kord.extensions.commands.converters.impl.optionalRole
-import com.kotlindiscord.kord.extensions.commands.converters.impl.optionalString
-import com.kotlindiscord.kord.extensions.commands.converters.impl.role
 import com.kotlindiscord.kord.extensions.commands.converters.impl.string
 import com.kotlindiscord.kord.extensions.components.forms.ModalForm
 import com.kotlindiscord.kord.extensions.extensions.Extension
 import com.kotlindiscord.kord.extensions.extensions.ephemeralSlashCommand
-import com.kotlindiscord.kord.extensions.extensions.event
 import com.kotlindiscord.kord.extensions.extensions.publicSlashCommand
 import com.kotlindiscord.kord.extensions.koin.KordExKoinComponent
 import com.kotlindiscord.kord.extensions.types.respond
 import com.kotlindiscord.kord.extensions.types.respondingPaginator
-import dev.kord.common.entity.ButtonStyle
-import dev.kord.common.entity.Snowflake
-import dev.kord.common.exception.RequestException
 import dev.kord.core.behavior.channel.createMessage
-import dev.kord.core.behavior.edit
-import dev.kord.core.behavior.interaction.respondEphemeral
-import dev.kord.core.behavior.interaction.response.respond
 import dev.kord.core.behavior.interaction.suggestString
-import dev.kord.core.behavior.interaction.updateEphemeralMessage
-import dev.kord.core.entity.Guild
 import dev.kord.core.entity.channel.TextChannel
 import dev.kord.core.entity.interaction.AutoCompleteInteraction
 import dev.kord.core.event.interaction.AutoCompleteInteractionCreateEvent
-import dev.kord.core.event.interaction.GuildButtonInteractionCreateEvent
 import dev.kord.core.exception.EntityNotFoundException
-import dev.kord.core.kordLogger
-import dev.kord.rest.builder.message.create.actionRow
 import dev.kord.rest.builder.message.create.embed
 import io.ktor.client.request.forms.*
 import io.ktor.utils.io.*
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.async
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.withContext
 import org.koin.core.component.get
 import org.koin.core.component.inject
 import org.koin.core.qualifier.named
-import java.time.Instant
 import java.time.ZoneId
 import java.time.ZonedDateTime
 import java.time.format.DateTimeFormatter
-import java.util.*
-
-enum class InteractionIDType(val prefix: String) {
-    POLL_VOTE_OPTION("p-vo#"),
-    POLL_FINISH_POLL_MENU("p-fpm#"),
-    POLL_FINISH_POLL_QUIETLY("p-fpq#"),
-    POLL_FINISH_POLL_LOUDLY("p-fpl#");
-
-    fun preppendTo(str: String) = "$prefix$str"
-
-    companion object {
-        fun takeFromStringOrNull(str: String): InteractionIDType? {
-            return entries.find {
-                str.startsWith(it.prefix)
-            }
-        }
-
-        fun removeFromString(str: String) = str.substringAfter('#')
-    }
-}
 
 class UtilsExtension<T : Any> : Extension(), KordExKoinComponent, SchedulerEventHandler<T> {
     override val name = "utils"
@@ -111,15 +71,6 @@ class UtilsExtension<T : Any> : Extension(), KordExKoinComponent, SchedulerEvent
     private lateinit var loggerChannel: TextChannel
 
     private val availableZoneIds = ZoneId.getAvailableZoneIds().toSortedSet()
-
-    // it might  be wiser to convert ULongs to RoleBehaviour and use their `.mention`
-    private fun ULong?.toMentionOn(guild: Guild): String? {
-        return when (this) {
-            null -> null
-            guild.id.value -> "@everyone"
-            else -> "<@&$this>"
-        }
-    }
 
     inner class StitchCoversArguments : Arguments() {
             val firstCover by long {
@@ -211,20 +162,6 @@ class UtilsExtension<T : Any> : Extension(), KordExKoinComponent, SchedulerEvent
         }
     }
 
-    inner class EditDateArguments : ScheduledPostArguments() {
-        val date by date {
-            name = "fecha"
-            description = "Nueva fecha."
-        }
-    }
-
-    inner class EditMentionRoleArguments : ScheduledPostArguments() {
-        val mention by role {
-            name = "mención"
-            description = "Rol al que mencionar."
-        }
-    }
-
     inner class WrittenItems : ModalForm() {
         override var title: String = "Aleatorizar."
 
@@ -232,49 +169,6 @@ class UtilsExtension<T : Any> : Extension(), KordExKoinComponent, SchedulerEvent
             label = "Un item por línea."
             maxLength = 900
             minLength = 3
-        }
-    }
-
-    inner class NewPollArguments : Arguments () {
-        val title by string {
-            name = "título"
-            description = "Título de la votación."
-            maxLength = 80
-        }
-
-        val pollOptionA by long {
-            name = "opción_a"
-            description = "Opción A."
-            autoCompleteCallback = titleAutoCompletionIdNamePairs
-        }
-
-        val pollOptionB by long {
-            description = "Opción B."
-            name = "opción_b"
-            autoCompleteCallback = titleAutoCompletionIdNamePairs
-        }
-
-        val pollOptionC by optionalLong {
-            description = "Opción C."
-            name = "opción_c"
-            autoCompleteCallback = titleAutoCompletionIdNamePairs
-        }
-
-        val pollOptionD by optionalLong {
-            description = "Opción D."
-            name = "opción_d"
-            autoCompleteCallback = titleAutoCompletionIdNamePairs
-        }
-
-        val description by optionalString {
-            description = "Descripción de la votación."
-            name = "descripción"
-            maxLength = 80
-        }
-
-        val singleVote by optionalBoolean {
-            name = "votoúnico"
-            description = "¿Sólo puede votarse una opción?"
         }
     }
 
@@ -570,245 +464,6 @@ class UtilsExtension<T : Any> : Extension(), KordExKoinComponent, SchedulerEvent
                         busyMutex.unlock()
                     } else {
                         content = "Lo siento, estoy ocupada procesando algo más."
-                    }
-                }
-            }
-        }
-
-        publicSlashCommand(::NewPollArguments) {
-            name = "votación"
-            description = "Crea una votación"
-            guild(config.guild)
-
-            action {
-                respond {
-                    val mangas = with (arguments) {
-                        listOfNotNull(pollOptionA, pollOptionB, pollOptionC, pollOptionD).distinct().let {
-                            db.getMangas(*it.toLongArray())
-                        }
-                    }
-
-                    val pollOptions = mangas.map { manga -> PollOption(manga.title) }
-
-                    @OptIn(EphemeralOrPublicView::class)
-                    if (mangas.size < 2) {
-                        respondWithError("No se encontraron suficientes entradas para crear una votación.")
-                        return@action
-                    }
-
-                    // if it's already busy let's not stitch the covers
-                    // I don't want the RPi die out of an OOM error or a busy and slow swap
-                    val imgBytes = if (busyMutex.tryLock()) try {
-                        withContext(Dispatchers.IO) {
-                            mangas.mapNotNull {
-                                it.imgURLSource ?: this@UtilsExtension.javaClass.getResource("no-cover.jpg")
-                                    .also { url ->
-                                        if (url === null)
-                                            kordLogger.warn {
-                                                "Default cover image for manga entries could not be loaded."
-                                            }
-                                    }
-                            }.takeIf { it.size > 1 }
-                                ?.let(::stitch)
-                                ?.asJpgByteArray()
-                        }
-                    } finally {
-                        busyMutex.unlock()
-                    } else null
-
-                    val pollCandidate = Poll(
-                        authorID = user.id.value.toLong(),
-                        title = arguments.title,
-                        description = arguments.description,
-                        imgByteArray = imgBytes,
-                        options = pollOptions,
-                        singleVote = arguments.singleVote ?: false,
-                    )
-
-                    val existingPoll = pollsDb.createPoll(pollCandidate)
-
-                    embeds.add(existingPoll.toEmbed())
-
-                    // voting choices:
-                    actionRow {
-                        existingPoll.options.forEachIndexed { i, opt ->
-                            interactionButton(
-                                ButtonStyle.Primary,
-                                InteractionIDType.POLL_VOTE_OPTION.preppendTo("${opt.id}"),
-                            ) {
-                                val icon = optionIcons.getOrNull(i)?.plus(" ") ?: ""
-
-                                label = "$icon${opt.description}".let {
-                                    if (it.length > 80) it.take(77) + "..." else it
-                                }
-                            }
-                        }
-                    }
-
-                    // config:
-                    actionRow {
-                        interactionButton(
-                            ButtonStyle.Danger,
-                            InteractionIDType.POLL_FINISH_POLL_MENU.preppendTo("${existingPoll.id}"),
-                        ) {
-                            label = "Cerrar"
-                        }
-                    }
-                }
-            }
-        }
-
-        event<GuildButtonInteractionCreateEvent> {
-            action {
-                val componentId = event.interaction.componentId
-                val id = InteractionIDType.removeFromString(componentId)
-
-                val requestType = InteractionIDType.takeFromStringOrNull(componentId)
-                    ?: return@action
-
-                when (requestType) {
-                    InteractionIDType.POLL_FINISH_POLL_MENU -> {
-                        event.interaction.respondEphemeral {
-                            content = "¿Cerrar con, o sin anuncios?"
-
-                            val pollIdPlusMsgId = "$id#${event.interaction.message.id.value}"
-
-                            actionRow {
-                                interactionButton(
-                                    ButtonStyle.Danger,
-                                    InteractionIDType.POLL_FINISH_POLL_LOUDLY.preppendTo(pollIdPlusMsgId),
-                                ) {
-                                    label = "Con anuncios"
-                                }
-
-                                interactionButton(
-                                    ButtonStyle.Danger,
-                                    InteractionIDType.POLL_FINISH_POLL_QUIETLY.preppendTo(pollIdPlusMsgId),
-                                ) {
-                                    label = "Sin anuncios"
-                                }
-                            }
-                        }
-                    }
-
-                    InteractionIDType.POLL_VOTE_OPTION -> {
-                        val deferredResponse = kord.async { event.interaction.deferEphemeralResponse() }
-
-                        val content: String = try {
-                            val optionID = UUID.fromString(id)
-                            val vote = pollsDb.vote(optionID, event.interaction.user.id.value)
-                            val votedPoll = pollsDb.getPollByOptionID(optionID)!!
-                            val pollEmbed = votedPoll.toEmbed()
-
-                            val optionDescription = votedPoll.options.first { it.id == optionID }.description
-
-                            event.interaction.message.edit {
-                                embeds = embeds?.apply { add(pollEmbed) } ?: mutableListOf(pollEmbed)
-                            }
-
-                            "Tu voto por **$optionDescription** fue ${if (vote) "registrado" else "revocado"}"
-                        } catch (e: Exception) {
-                            when (e) {
-                                is IllegalArgumentException,
-                                is PollException -> {
-                                    kordLogger.error(e) { "Exception during voting process. Probably a mistake coming from the front-end." }
-                                    "Lo siento, hubo un error con la votación."
-                                }
-                                else -> {
-                                    deferredResponse.await().respond {
-                                        content = "Lo siento, ocurrió un error desconocido durante la votación."
-                                    }
-                                    throw e
-                                }
-                            }
-                        }
-
-                        deferredResponse.await().respond {
-                            this.content = content
-                        }
-                    }
-
-                    InteractionIDType.POLL_FINISH_POLL_QUIETLY, InteractionIDType.POLL_FINISH_POLL_LOUDLY -> {
-                        val memberIsHelper = config.allowedRole in event.interaction.user.asMember().roleIds
-
-                        if (!memberIsHelper) {
-                            event.interaction.updateEphemeralMessage {
-                                content = "Lo siento, no tienes autorización para cerrar esta votación"
-                            }
-
-                            return@action
-                        }
-
-                        val quietly = requestType == InteractionIDType.POLL_FINISH_POLL_QUIETLY
-
-                        val (pollID, pollMessageID) = id.split('#').let { (a, b) ->
-                            a.toLong() to Snowflake(b.toLong())
-                        }
-
-                        try {
-                            pollsDb.finishPoll(pollID)
-                        } catch (e: PollUnsuccessfulOperationException) {
-                            event.interaction.updateEphemeralMessage {
-                                components = mutableListOf()
-
-                                content = "No pude cerrar la votación. ¿Puede que ya esté cerrada?"
-                            }
-
-                            return@action
-                        }
-
-                        val finishedPoll = pollsDb.getPoll(pollID)
-
-                        if (finishedPoll !== null && !quietly) {
-                            val rankedOptions = finishedPoll.options.groupBy { it.votes }
-                                .toSortedMap(compareByDescending { it })
-                                .values.flatMapIndexed { rank, pollOptions ->
-                                    pollOptions.map { o -> rank to o }
-                                }
-
-                            val resultsText = rankedOptions.joinToString(
-                                separator="\n",
-                                prefix="«${finishedPoll.title}»: votación finalizada\n\n"
-                            ) { (rank, option) ->
-                                "${if (rank == 0) "\uD83C\uDFC6 " else ""}${option.description} (${option.votes})"
-                            }
-
-                            try {
-                                scheduler.schedule(
-                                    DiscordWebhookMessage(
-                                        content=config.announcementRole.value.toMentionOn(kord.getGuildOrThrow(config.guild)),
-                                        embedText=resultsText,
-                                        execInstant=Instant.now(),
-                                    )
-                                )
-                            } catch (e: RequestException) {
-                                kordLogger.error(e) { "There was a problem trying to resolve a guild for a scheduled message" }
-                            }
-
-                            scheduler.schedule(
-                                XPost(
-                                    text=resultsText,
-                                    execInstant=Instant.now(),
-                                )
-                            )
-                        }
-
-                        kord.rest.channel.editMessage(event.interaction.channelId, pollMessageID) {
-                            finishedPoll?.toEmbed()?.let { pollEmbed ->
-                                embeds = embeds?.apply { add(pollEmbed) } ?: mutableListOf(pollEmbed)
-                            }
-                            components = mutableListOf()
-                        }
-
-                        event.interaction.updateEphemeralMessage {
-                            components = mutableListOf()
-
-                            content = if (requestType == InteractionIDType.POLL_FINISH_POLL_LOUDLY) {
-                                "Cerrado y anunciado."
-                            } else {
-                                "Cerrado silenciosamente."
-                            }
-                        }
                     }
                 }
             }
